Update README.md
This a guessing game program that generates a random number and challenges the user to guess it. The program prompt the user to input their guess, compare it to the generated number, and provide feedback if the guess is too high or too low. This continues until the user correctly guesses the number and then display the number of attempts it took to win the game.
#Guessing game
#prodigyInfoTech